@@ -26,7 +26,7 @@ dependencies = [
     "psycopg2-binary>=2.9.11",
     "pgvector>=0.4.2",
 ]
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 readme = "README.md"
 license = {text = "MIT"}
 
@@ -60,7 +60,7 @@ dev = [
 
 [tool.ruff]
 line-length = 100
-target-version = "py310"
+target-version = "py311"
 extend-exclude = ["build", "dist", ".venv"]
 
 [tool.ruff.format]
@@ -86,7 +86,7 @@ select = [
     "ARG",
     "RUF",
 ]
-ignore = ["E501", "I001", "UP045"]
+ignore = ["E501", "I001", "UP045", "SIM105"]
 
 [tool.ruff.lint.isort]
 combine-as-imports = true
